// libs/modules/sps-rbac/models/identity/frontend/component/variants/sps-lite/admin-select-input/src/lib/interface.ts
import { IModel } from "@sps/sps-rbac/models/identity/contracts/root";
import { IModel as IModelExtended } from "@sps/sps-rbac/models/identity/contracts/extended";
import { UseFormReturn } from "react-hook-form";
import { ISpsComponentBase } from "@sps/ui-adapter";

export const variant = "admin-select-input" as const;

export interface IComponentProps extends ISpsComponentBase {
  variant: typeof variant;
  formFieldName: string;
  form: UseFormReturn<any>;
  renderField?: keyof IModel;
  className?: string;
}

export interface IComponentPropsExtended extends IComponentProps {
  data: IModelExtended[];
}